moving tfslim_classifiers module up one-level; adding default deployment support
