+++ eta/classifiers/__init__.py
@@ -8,7 +8,7 @@ Brian Moore, brian@voxel51.com
 '''
 
 # Import all classifiers into the `eta.classifiers` namespace
-from .tfslim.tfslim_classifiers import TFSlimClassifier, TFSlimClassifierConfig
+from .tfslim_classifiers import TFSlimClassifier, TFSlimClassifierConfig
 from .vgg16_classifiers import VGG16Classifier, VGG16ClassifierConfig
 from .voting_classifiers import VideoFramesVotingClassifier, \
                                 VideoFramesVotingClassifierConfig
