--- eta/classifiers/tfslim/__init__.py
@@ -1,8 +1,0 @@
-'''
-Package declaration.
-
-Copyright 2017-2019, Voxel51, Inc.
-voxel51.com
-
-Brian Moore, brian@voxel51.com
-'''
